feat: implementar logica completa de Reserva y Mesa (crear, cancelar, cambiar estado, liberar mesa automaticamente)

+++ backend/src/main/java/com/lsvf/backend/entities/Mesa.java
@@ -22,7 +22,7 @@ public class Mesa {
     private Integer numeroMesa;
 
     @Column(nullable = false)
-    private Long capacidad;
+    private Integer capacidad;
 
     @Column(nullable = false)
     @Enumerated(EnumType.STRING)
